Move `util` submodule up one level
