+++ curated_transformers/models/bert/encoder.py
@@ -10,8 +10,8 @@ from .layer import BertEncoderLayer
 from ..attention import AttentionMask
 from ._hf import convert_hf_config, convert_hf_state_dict
 from ..output import ModelOutput
-from ..util.serde import DeserializationParamBucket
-from ..util.hf import _param_buckets_for_bert_qkv
+from ...util.serde import DeserializationParamBucket
+from ...util.hf import _param_buckets_for_bert_qkv
 
 
 # Only provided as typing.Self in Python 3.11+.
