+++ curated_transformers/models/gpt_neox/decoder.py
@@ -11,7 +11,7 @@ from ..output import ModelOutputWithCache
 from ._hf import convert_hf_config, convert_hf_state_dict
 from .config import GPTNeoXConfig
 from .layer import GPTNeoXDecoderLayer
-from ..util.serde import DeserializationParamBucket
+from ...util.serde import DeserializationParamBucket
 
 
 # Only provided as typing.Self in Python 3.11+.
