+++ curated_transformers/models/hf_hub.py
@@ -6,7 +6,7 @@ from requests import HTTPError  # type: ignore
 import torch
 from torch import Tensor
 
-from .util.serde import load_model_from_checkpoints, DeserializationParamBucket
+from ..util.serde import load_model_from_checkpoints, DeserializationParamBucket
 
 HF_MODEL_CONFIG = "config.json"
 HF_MODEL_CHECKPOINT = "pytorch_model.bin"
